+++ ci_target/xss.js
@@ -1,3 +1,5 @@
-
-var name = getUrlParam('name');
-document.write('<div>' + name + '</div>');
+// XSS test case
+// document.write with location.search
+var name = location.search;
+document.write(name);
+// end of file